refactor(web): merge LaunchScreenView into LaunchScreen for naming consistency (WA-2558)
Review feedback: the folder is LaunchScreen but exposed two component names
(LaunchScreen + LaunchScreenView), and the container/presentational *View split
isn't a pattern used elsewhere in the repo. Collapse into a single LaunchScreen
component; rename the story/test to LaunchScreen.*. Timing logic stays in
useLaunchScreen (unit-tested); the component test mocks that hook and the story
uses withMockProvider({ shadcn: true }) with the visual snapshot disabled.
