+++ apps/web/src/components/common/LaunchScreen/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState, type ReactElement } from 'react'
+import { cn } from '@/utils/cn'
 import { useLaunchScreen } from './useLaunchScreen'
-import LaunchScreenView from './LaunchScreenView'
+import css from './LaunchScreen.module.css'
 
-// Decorative, time-driven steps — boot has no granular progress signal, so this is purely cosmetic.
 const LAUNCH_STEPS = [
   { progress: 30, caption: 'Loading your workspace…' },
   { progress: 65, caption: 'Fetching your accounts…' },
@@ -13,25 +13,24 @@ const STEP_INTERVAL_MS = 700
 const EXIT_DURATION_MS = 300
 
 /**
- * App-boot launch screen. Renders a full-screen branded splash from first mount until the
- * app shell is ready (see {@link useLaunchScreen}), then fades out and unmounts. Mounted once
- * in `_app`, so it never replays on client-side navigations.
+ * App-boot launch screen: a full-screen branded splash shown from first mount until the app
+ * shell is ready (see {@link useLaunchScreen}), then faded out and unmounted. Mounted once in
+ * `_app`, so it never replays on client-side navigations.
  */
 function LaunchScreen(): ReactElement | null {
   const { visible } = useLaunchScreen()
   const [rendered, setRendered] = useState(true)
   const [stepIndex, setStepIndex] = useState(0)
 
-  // Advance the decorative steps while visible.
   useEffect(() => {
     if (!visible) return
-    const id = setInterval(() => {
-      setStepIndex((index) => Math.min(index + 1, LAUNCH_STEPS.length - 1))
-    }, STEP_INTERVAL_MS)
+    const id = setInterval(
+      () => setStepIndex((index) => Math.min(index + 1, LAUNCH_STEPS.length - 1)),
+      STEP_INTERVAL_MS,
+    )
     return () => clearInterval(id)
   }, [visible])
 
-  // Keep mounted through the fade-out, then unmount.
   useEffect(() => {
     if (visible) return
     const id = setTimeout(() => setRendered(false), EXIT_DURATION_MS)
@@ -40,9 +39,41 @@ function LaunchScreen(): ReactElement | null {
 
   if (!rendered) return null
 
-  const step = LAUNCH_STEPS[stepIndex]
-  // Snap to 100% during the exit fade so the bar reads as complete.
-  return <LaunchScreenView progress={visible ? step.progress : 100} caption={step.caption} exiting={!visible} />
+  const exiting = !visible
+  const { progress, caption } = LAUNCH_STEPS[stepIndex]
+
+  return (
+    <div
+      role="status"
+      aria-busy={!exiting}
+      aria-live="polite"
+      aria-label="Loading Safe{Wallet}"
+      data-testid="launch-screen"
+      className={cn(
+        'fixed inset-0 z-[1401] flex flex-col items-center justify-center gap-8 bg-background transition-opacity duration-300',
+        exiting && 'pointer-events-none opacity-0',
+      )}
+    >
+      <div className="relative flex items-center justify-center">
+        <span aria-hidden className={cn('absolute size-40 rounded-full', css.halo)} />
+        <div className={css.breathe}>
+          <img src="/images/logo-no-text.svg" alt="Safe" width={72} height={72} className="size-[72px] dark:hidden" />
+          <span aria-hidden className={cn('hidden size-[72px] dark:block', css.logoDarkFill)} />
+        </div>
+      </div>
+
+      <div className="flex flex-col items-center gap-4">
+        <div className="h-1 w-40 overflow-hidden rounded-full bg-secondary">
+          <div
+            data-testid="launch-progress-bar"
+            className={cn('h-full rounded-full', css.bar)}
+            style={{ width: `${exiting ? 100 : progress}%`, backgroundColor: 'var(--color-static-text-brand)' }}
+          />
+        </div>
+        <p className="min-h-5 text-sm text-muted-foreground">{caption}</p>
+      </div>
+    </div>
+  )
 }
 
 export default LaunchScreen
